requirements.txt
# FX LLM Bot - Requirements
# Python 3.11+

# ========== Core ==========
MetaTrader5>=5.0.45
pandas>=2.1.0
numpy>=1.25.0
pandas-ta>=0.3.14b

# ========== Backtesting ==========
vectorbt>=0.26.0

# ========== LLM ==========
openai>=1.12.0

# ========== Scheduling ==========
APScheduler>=3.10.4

# ========== Config & Data ==========
PyYAML>=6.0.1
pytz>=2024.1

# ========== Monitoring ==========
streamlit>=1.31.0
plotly>=5.18.0

# ========== Notifications ==========
requests>=2.31.0
aiohttp>=3.9.0

# ========== Database ==========
# SQLite is built-in

# ========== Dev/Test ==========
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
